third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/width-expected.txt

This is a testharness.js-based test.
Found 3 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'width' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
[FAIL] Can set 'min-width' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
[FAIL] Can set 'max-width' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
Harness: the test ran to completion.
